pyproject.toml
[project]
authors = [
    {name = "Mozilla", email = "conduit-team@mozilla.com"},
]
classifiers = [
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
description =  "Unified Lando Project"
dependencies = [
    "celery",
    "cryptography",
    "datadog",
    "django ~= 5.0",
    "django-libsass",
    "django-ninja",
    "django-storages[google]",
    "django_compressor",
    "jinja2",
    "kombu",
    "mercurial==6.1.4",
    "mots",
    "mozilla_django_oidc",
    "networkx",
    "psycopg2-binary",
    "python-hglib==2.6.2",
    "python-jose",
    "redis",
    "rs-parsepatch",
    "sentry-sdk",
    "setuptools-scm",
    "simple-github",
    "uwsgi",
]
name = "lando"
requires-python = ">=3.10"
dynamic = ["version"]

[project.optional-dependencies]
code-quality = ["black", "ruff"]
testing = [
  "pytest",
  "pytest-cov",
  "pytest-django",
  "requests-mock"
]

[project.scripts]
lando = "lando.manage:main"

[build-system]
requires = [
    "setuptools>=64",
    "setuptools_scm>=8",
    "pip-tools",
    "wheel",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_file = "src/lando/version.py"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "lando.test_settings"
addopts = "--cov --cov-report html"

testpaths = [
    "src/lando/api",
    "src/lando/dockerflow",
    "src/lando/main/tests",
    "src/lando/tests",
    "src/lando/ui/tests",
]
